[REFACTOR] refactor component calling, fix cycle dependencies
